requirements.txt
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colored==2.3.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.1.2
mypy==1.15.0
mypy-extensions==1.0.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
types-jsonschema==4.23.0.20241208
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
validators==0.34.0
